+++ bootmgr-rs-slint/src/ui/slint_inc.rs
@@ -5,12 +5,12 @@
 //!
 //! This primarily provides the `Ui` export.
 
-#![allow(
-    clippy::missing_docs_in_private_items,
-    reason = "Slint auto generated files produce false lint warnings by the hundreds"
-)]
-#![allow(
-    clippy::unwrap_used,
-    reason = "Slint auto generated files produce false lint warnings by the hundreds"
-)]
+// The code included in include_modules! is completely auto generated. As a result, it may erroneously cause lint warnings since
+// the auto generated code may freely use unwrap or have missing documentation. Therefore, every single clippy lint is disabled
+// in this specific module.
+#![allow(clippy::all)]
+#![allow(clippy::pedantic)]
+#![allow(clippy::nursery)]
+#![allow(clippy::restriction)]
+
 slint::include_modules!();